requirements/requirements.txt
boto==2.26.1
mongoengine==0.10.6
dropbox==6.8.0
django-bootstrap3==8.2.1
pymongo==3.0.2